+++ app/team/page.js
@@ -7,7 +7,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-
+import teamObject from "@/data/Team/teamObject";
 import {
   Carousel,
   CarouselContent,
@@ -54,384 +54,7 @@ function CarouselDemo() {
   );
 }
 const page = () => {
-  const teamDataObject = {
-    Leads: [
-      {
-        name: "Vansh Baghel",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Change is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-    ],
-    Technical: [
-      {
-        name: "Vansh Baghel",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-    ],
-    Secretariat: [
-      {
-        name: "Vansh Baghel",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-    ],
-    Logistics: [
-      {
-        name: "Vansh Baghel",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-    ],
-
-    Event: [
-      {
-        name: "Vansh Baghel",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-    ],
-
-    Marketing: [
-      {
-        name: "Vansh Baghel",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-    ],
-
-    Press: [
-      {
-        name: "Vansh Baghel",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-    ],
-
-    Outreach: [
-      {
-        name: "Vansh Baghel",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Growth is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-      {
-        name: "Ritik Jain",
-        role: "Lead",
-        quote: "Change is painful, but what is worse is staying the same.",
-        socials: [
-          {
-            icon: "Team/Vectorgithub.svg",
-            link: "#",
-          },
-          {
-            icon: "Team/Vectorinstagram.svg",
-            link: "#",
-          },
-        ],
-      },
-    ],
-  };
+  const teamDataObject = teamObject;
   return (
     <div className="flex flex-col mb-0">
       <div className="relative mt-0 sm:mt-[120px] sm:mb-[200px]">
