+++ data/Team/teamObject.js
@@ -1,0 +1,380 @@
+const teamDataObject = {
+  Leads: [
+    {
+      name: "Vansh Baghel",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Change is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+  ],
+  Technical: [
+    {
+      name: "Vansh Baghel",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+  ],
+  Secretariat: [
+    {
+      name: "Vansh Baghel",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+  ],
+  Logistics: [
+    {
+      name: "Vansh Baghel",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+  ],
+
+  Event: [
+    {
+      name: "Vansh Baghel",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+  ],
+
+  Marketing: [
+    {
+      name: "Vansh Baghel",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+  ],
+
+  Press: [
+    {
+      name: "Vansh Baghel",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+  ],
+
+  Outreach: [
+    {
+      name: "Vansh Baghel",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Growth is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+    {
+      name: "Ritik Jain",
+      role: "Lead",
+      quote: "Change is painful, but what is worse is staying the same.",
+      socials: [
+        {
+          icon: "Team/Vectorgithub.svg",
+          link: "#",
+        },
+        {
+          icon: "Team/Vectorinstagram.svg",
+          link: "#",
+        },
+      ],
+    },
+  ],
+};
+
+export default teamDataObject;
